forge_script

Module execute

Structs

ExecutedState

State after the script has been executed.

ExecutionArtifacts

Container for data being collected after execution.

ExecutionData

Container for data we need for execution which can only be obtained after linking stage.

LinkedState

State after linking, contains the linked build data along with library addresses and optional array of libraries that need to be predeployed.

PreExecutionState

Same as LinkedState, but also contains ExecutionData.

RpcData

Container for information about RPC-endpoints used during script execution.