Struct forge_script::execute::RpcData

pub struct RpcData {
    pub total_rpcs: HashSet<String>,
    pub missing_rpc: bool,
}

Container for information about RPC-endpoints used during script execution.

Fields
total_rpcs: HashSet<String>

Unique list of rpc urls present.

missing_rpc: bool

If true, one of the transactions did not have a rpc.

Implementations

impl RpcData

fn from_transactions(txs: &BroadcastableTransactions) -> Self

Iterates over script transactions and collects RPC urls.

pub fn is_multi_chain(&self) -> bool

Returns true if script might be multi-chain. Returns false positive in case when missing rpc is the same as the only rpc present.

async fn check_shanghai_support(&self) -> Result<()>

Checks if all RPCs support EIP-3855. Prints a warning if not.