Function forge_script::receipts::clear_pendings

pub async fn clear_pendings(
    provider: Arc<RetryProvider>,
    deployment_sequence: &mut ScriptSequence,
    tx_hashes: Option<Vec<TxHash>>
) -> Result<()>

Traverses a set of pendings and either finds receipts, or clears them from the deployment sequence.

If no tx_hashes are provided, then deployment_sequence.pending will be used. For each tx_hash, we check if it has confirmed. If it has confirmed, we push the receipt (if successful) or push an error (if revert). If the transaction has not confirmed, but can be found in the node’s mempool, we wait for its receipt to be available. If the transaction has not confirmed, and cannot be found in the mempool, we remove it from the deploy_sequence.pending vector so that it will be rebroadcast in later steps.