forge_script

Module sequence

Structs

ScriptSequence

Helper that saves the transactions sequence and its state on which transactions have been broadcasted

SensitiveScriptSequence

Sensitive info from the script sequence which is saved into the cache folder

SensitiveTransactionMetadata

Sensitive values from the transactions in a script sequence

Enums

ScriptSequenceKind

Constants

DRY_RUN_DIR

Functions

get_commit_hash

Returns the commit hash of the project if it exists

sig_to_file_name

Converts the sig argument into the corresponding file path.