forge_script

Module build

Structs

BuildData

Container for the compiled contracts.

CompiledState

State after we have determined and compiled target contract to be executed.

LinkedBuildData

Container for the linked contracts and their dependencies

PreprocessedState

First state basically containing only inputs of the user.

Enums

ScriptPredeployLibraries